requirements.txt
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langchain_pinecone
pinecone_client
streamlit
pypdf
python-dotenv
tiktoken
redis